# docs/api-ui/openapi.yaml
openapi: 3.1.0
info:
  title: rsyslox API
  version: "0.4.0"
  description: |
    REST API for rsyslox — a syslog viewer and management tool.

    ## Authentication

    Most endpoints require authentication via one of two methods:

    | Method | Header | Access level |
    |--------|--------|--------------|
    | Admin session token | `X-Session-Token: <token>` | Full access |
    | Read-only API key | `X-API-Key: <key>` | `/api/logs`, `/api/meta` only |

    Obtain an admin session token via `POST /api/admin/login`.
servers:
  - url: http://localhost:8000
    description: Local development
...
paths:
...
  /api/meta:
    get:
      tags: [meta]
      summary: List available columns
      operationId: getMeta
      security:
        - SessionToken: []
        - ApiKey: []
      responses:
        "200":
          description: Available column names
          content:
            application/json:
              schema: { $ref: "#/components/schemas/MetaResponse" }
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/meta/{column}:
    get:
      tags: [meta]
      summary: Get distinct values for a column
      operationId: getMetaColumn
      security:
        - SessionToken: []
        - ApiKey: []
      parameters:
        - name: column
          in: path
          required: true
          description: "Column name (e.g. `FromHost`, `SysLogTag`, `Severity`)"
          schema: { type: string }
        - $ref: "#/components/parameters/StartDate"
        - $ref: "#/components/parameters/EndDate"
        - name: Severity
          in: query
          schema: { type: integer }
        - name: Facility
          in: query
          schema: { type: integer }
        - name: FromHost
          in: query
          schema: { type: string }
        - name: SysLogTag
          in: query
          schema: { type: string }
        - name: Message
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Distinct values for the column
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { type: string }
                    description: String column values
                  - type: array
                    items: { type: integer }
                    description: Integer column values
                  - type: array
                    items: { $ref: "#/components/schemas/MetaValue" }
                    description: Severity or Facility values with labels
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Admin: auth ───────────────────────────────────────────────────────────
...
  /api/admin/config:
    get:
      tags: [admin]
      summary: Get current configuration
      operationId: getConfig
      security:
        - SessionToken: []
      responses:
        "200":
          description: Current configuration (passwords excluded)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigView" }
        "401":
          $ref: "#/components/responses/Unauthorized"

    patch:
      tags: [admin]
      summary: Update configuration
      operationId: updateConfig
      security:
        - SessionToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConfigUpdateRequest" }
      responses:
        "200":
          description: Updated configuration
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ConfigView" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"

  # ── Admin: keys ───────────────────────────────────────────────────────────
  /api/admin/keys:
    get:
      tags: [admin]
      summary: List read-only API keys
      operationId: listKeys
      security:
        - SessionToken: []
      responses:
        "200":
          description: List of key names (hashes never returned)
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/KeyInfo" }
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [admin]
      summary: Create a new read-only API key
      operationId: createKey
      security:
        - SessionToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  description: Unique human-readable name for the key
      responses:
        "201":
          description: Key created — plaintext returned exactly once
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CreateKeyResponse" }
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Key name already exists
          content:
            application/json:
              schema: { $ref: "#/components/schemas/APIError" }
...
components:
  securitySchemes:
    SessionToken:
      type: apiKey
      in: header
      name: X-Session-Token
      description: Admin session token obtained from POST /api/admin/login
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: Read-only API key (access to /api/logs and /api/meta only)
  parameters:
...
    StartDate:
      name: start_date
      in: query
      description: "Start of time range (format: `2006-01-02 15:04:05`)"
      schema: { type: string, example: "2025-01-01 00:00:00" }
    EndDate:
      name: end_date
      in: query
      description: "End of time range (format: `2006-01-02 15:04:05`)"
      schema: { type: string, example: "2025-12-31 23:59:59" }
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema: { $ref: "#/components/schemas/APIError" }
  schemas:
    APIError:
      type: object
      properties:
        code:    { type: string, example: "INVALID_PARAMETER" }
        message: { type: string, example: "limit must be between 1 and 1000" }
        details: { type: string }
        field:   { type: string }
...
    MetaValue:
      type: object
      properties:
        val:   { type: integer }
        label: { type: string, example: "Warning" }
    MetaResponse:
      type: object
      properties:
        available_columns:
          type: array
          items: { type: string }
        usage: { type: string }
...
    ConfigView:
      type: object
      properties:
        server:
          type: object
          properties:
            host:                  { type: string }
            port:                  { type: integer }
            use_ssl:               { type: boolean }
            ssl_cert:              { type: string }
            ssl_key:               { type: string }
            allowed_origins:       { type: array, items: { type: string } }
            auto_refresh_interval: { type: integer }
        database:
          type: object
          properties:
            host: { type: string }
            port: { type: integer }
            name: { type: string }
            user: { type: string }
        cleanup:
          type: object
          properties:
            enabled:           { type: boolean }
            disk_path:         { type: string }
            threshold_percent: { type: number }
            batch_size:        { type: integer }
            interval_seconds:  { type: integer }
    ConfigUpdateRequest:
      type: object
      properties:
        server:
          type: object
          properties:
            allowed_origins:       { type: array, items: { type: string } }
            auto_refresh_interval: { type: integer, minimum: 5 }
            use_ssl:               { type: boolean }
        cleanup:
          type: object
          properties:
            enabled:           { type: boolean }
            disk_path:         { type: string }
            threshold_percent: { type: number, minimum: 1, maximum: 100 }
            batch_size:        { type: integer, minimum: 1 }
            interval_seconds:  { type: integer, minimum: 60 }
    KeyInfo:
      type: object
      properties:
        name: { type: string }
    CreateKeyResponse:
      type: object
      properties:
        name:    { type: string }
        key:     { type: string, description: "Plaintext key — shown exactly once" }
        message: { type: string }
...